requirements.txt
discord.py
requests
pillow
flask
